requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
pandas==2.1.3
matplotlib==3.8.2
pytest==7.4.3
black==23.11.0
mypy==1.7.1
isort==5.12.0
twine==4.0.2
setuptools==68.2.2
wheel==0.41.3 